moved scenetracing integration into the SDK
