+++ package-dev/Runtime/SentryIntegrations.cs
@@ -1,7 +1,3 @@
-#if UNITY_2020_3_OR_NEWER
-#define SENTRY_SCENE_MANAGER_TRACING_INTEGRATION
-#endif
-
 #if !UNITY_EDITOR
 #if UNITY_WEBGL
 #define SENTRY_WEBGL
@@ -28,20 +24,7 @@ namespace Sentry.Unity
                     options.AddIntegration(new StartupTracingIntegration());
                 }
 #endif
-#if SENTRY_SCENE_MANAGER_TRACING_INTEGRATION
-                if (options.AutoSceneLoadTraces)
-                {
-                    options.AddIntegration(new SceneManagerTracingIntegration());
-                }
-#endif
             }
-            else
-            {
-#if SENTRY_SCENE_MANAGER_TRACING_INTEGRATION
-                options.DiagnosticLogger?.LogDebug("Skipping SceneManagerTracing integration because performance tracing is disabled.");
-#endif
-            }
-
         }
     }
 
@@ -133,65 +116,6 @@ namespace Sentry.Unity
             });
 
             StartupAlreadyCaptured = true;
-        }
-    }
-#endif
-
-#if SENTRY_SCENE_MANAGER_TRACING_INTEGRATION
-    public class SceneManagerTracingIntegration : ISdkIntegration
-    {
-        private static IDiagnosticLogger Logger;
-
-        public void Register(IHub hub, SentryOptions options)
-        {
-            Logger = options.DiagnosticLogger;
-
-            if (SceneManagerAPI.overrideAPI != null)
-            {
-                // TODO: Add a place to put a custom 'SceneManagerAPI' on the editor window so we can "decorate" it.
-                Logger?.LogWarning("Registering SceneManagerTracing integration - overwriting the previous SceneManagerAPI.overrideAPI.");
-            }
-
-            SceneManagerAPI.overrideAPI = new SceneManagerTracingAPI(Logger);
-        }
-    }
-
-    public class SceneManagerTracingAPI : SceneManagerAPI
-    {
-        public const string TransactionOperation = "scene.load";
-        private const string SpanOperation = "scene.load";
-        private readonly IDiagnosticLogger _logger;
-
-        public SceneManagerTracingAPI(IDiagnosticLogger logger)
-        {
-            _logger = logger;
-        }
-
-        protected override AsyncOperation LoadSceneAsyncByNameOrIndex(string sceneName, int sceneBuildIndex, LoadSceneParameters parameters, bool mustCompleteNextFrame)
-        {
-            _logger?.LogInfo("Creating '{0}' transaction for '{1}'.", TransactionOperation, sceneName);
-
-            var transaction = SentrySdk.StartTransaction("scene.loading", TransactionOperation);
-            SentrySdk.ConfigureScope(scope => scope.Transaction = transaction);
-
-            _logger?.LogDebug("Creating '{0}' span.", SpanOperation);
-            var span = SentrySdk.GetSpan()?.StartChild(SpanOperation, sceneName ?? $"buildIndex:{sceneBuildIndex}");
-
-            var asyncOp = base.LoadSceneAsyncByNameOrIndex(sceneName, sceneBuildIndex, parameters, mustCompleteNextFrame);
-
-            // TODO: setExtra()? e.g. from the LoadSceneParameters:
-            // https://github.com/Unity-Technologies/UnityCsReference/blob/02d565cf3dd0f6b15069ba976064c75dc2705b08/Runtime/Export/SceneManager/SceneManager.cs#L30
-            // Note: asyncOp.completed triggers in the next frame after finishing (so the time isn't precise).
-            // https://docs.unity3d.com/2020.3/Documentation/ScriptReference/AsyncOperation-completed.html
-            asyncOp.completed += _ =>
-            {
-                _logger?.LogInfo("Finishing '{0}' transaction for '{1}'.", TransactionOperation, sceneName);
-
-                span?.Finish(SpanStatus.Ok);
-                transaction.Finish(SpanStatus.Ok);
-            };
-
-            return asyncOp;
         }
     }
 #endif
